+++ kod/visionlabfotopasca/.gitignore
