+++ kod/visionlabfotopasca/Main.py
