+++ kod/visionlabfotopasca/Src/GUI/Alarm.py
