+++ kod/visionlabfotopasca/Src/GUI/CVGui/CVButton.py
