+++ kod/visionlabfotopasca/Src/GUI/CVGui/CVRoi.py
