+++ kod/visionlabfotopasca/Src/GUI/CVGui/CVSlider.py
