+++ kod/visionlabfotopasca/Src/GUI/CVGui/CVWindow.py
