+++ kod/visionlabfotopasca/Src/GUI/CVGui/RoiWindow.py
