+++ kod/visionlabfotopasca/Src/GUI/CVGui/RuntimeWindow.py
