+++ kod/visionlabfotopasca/Src/GUI/GUI.py
