+++ kod/visionlabfotopasca/Src/GUI/TKGui/ConfigurationWindow.py
