+++ kod/visionlabfotopasca/Src/Process/eventHandler.py
