+++ kod/visionlabfotopasca/res/conf.obj
